created new CSV for every subproblem, and implemented new Knns for different tasks
