+++ models/knn/basicknn.py
@@ -1,0 +1,9 @@
+
+from random import seed
+from random import randrange
+from csv import reader
+from math import sqrt
+
+
+def readfile (filename):
+    df = pd
